# api_specification.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Solar-Eye API",
    "version": "1.0.0",
    "description": "AI 기반 태양광 패널 결함 탐지 및 유지보수 솔루션 API 명세서"
  },
  "servers": [
    {
      "url": "https://api.solar-eye.com",
      "description": "Production Server"
    },
    {
      "url": "http://localhost:8000",
      "description": "Development Server"
    }
  ],
  "paths": {
...
    "/api/v1/reports/daily": {
      "get": {
        "tags": ["Reports"],
        "summary": "일간 리포트 조회",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "panel_id",
            "in": "query",
            "required": true,
            "schema": { "type": "integer" },
            "description": "패널 ID"
          },
          {
            "name": "date",
            "in": "query",
            "schema": { "type": "string", "format": "date" },
            "description": "조회 날짜 (기본값: 오늘)"
          }
        ],
        "responses": {
          "200": {
            "description": "일간 리포트",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": { "$ref": "#/components/schemas/DailyReport" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/weekly": {
      "get": {
        "tags": ["Reports"],
        "summary": "주간 리포트 조회",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "panel_id",
            "in": "query",
            "required": true,
            "schema": { "type": "integer" }
          },
          {
            "name": "start_date",
            "in": "query",
            "schema": { "type": "string", "format": "date" },
            "description": "주 시작일"
          }
        ],
        "responses": {
          "200": {
            "description": "주간 리포트",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": { "$ref": "#/components/schemas/WeeklyReport" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/economic": {
      "get": {
        "tags": ["Reports"],
        "summary": "경제성 분석 리포트",
        "description": "오염으로 인한 발전 손실과 청소 비용 대비 효과를 분석합니다.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "panel_id",
            "in": "query",
            "required": true,
            "schema": { "type": "integer" }
          },
          {
            "name": "period",
            "in": "query",
            "schema": { "type": "string", "enum": ["week", "month", "quarter"], "default": "month" },
            "description": "분석 기간"
          }
        ],
        "responses": {
          "200": {
            "description": "경제성 분석 리포트",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": { "$ref": "#/components/schemas/EconomicReport" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Firebase ID Token",
        "description": "Firebase Authentication ID Token"
      }
    },
    "schemas": {
...
      "DailyReport": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "panel_id": { "type": "integer", "example": 1 },
          "report_date": { "type": "string", "format": "date", "example": "2026-01-16" },
          "total_frames_analyzed": { "type": "integer", "example": 8640 },
          "total_defects": { "type": "integer", "example": 0 },
          "total_soiling": { "type": "integer", "example": 5 },
          "defect_breakdown": {
            "type": "object",
            "properties": {
              "broken": { "type": "integer", "example": 0 },
              "crack": { "type": "integer", "example": 0 },
              "hotspot": { "type": "integer", "example": 0 }
            }
          },
          "soiling_breakdown": {
            "type": "object",
            "properties": {
              "dust": { "type": "integer", "example": 3 },
              "bird_drop": { "type": "integer", "example": 2 },
              "leaves": { "type": "integer", "example": 0 }
            }
          },
          "estimated_efficiency_loss": { "type": "number", "example": 2.5, "description": "추정 효율 손실률 (%)" },
          "estimated_daily_loss": { "type": "number", "example": 1500, "description": "추정 일일 손실액 (원)" },
          "hourly_stats": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "hour": { "type": "integer", "example": 14 },
                "defect_count": { "type": "integer", "example": 0 },
                "soiling_count": { "type": "integer", "example": 1 }
              }
            }
          },
          "weather_summary": {
            "type": "object",
            "properties": {
              "avg_temperature": { "type": "number", "example": 5.2 },
              "precipitation": { "type": "boolean", "example": false },
              "condition": { "type": "string", "example": "맑음" }
            }
          },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "WeeklyReport": {
        "type": "object",
        "properties": {
          "panel_id": { "type": "integer", "example": 1 },
          "start_date": { "type": "string", "format": "date", "example": "2026-01-10" },
          "end_date": { "type": "string", "format": "date", "example": "2026-01-16" },
          "total_defects": { "type": "integer", "example": 1 },
          "total_soiling": { "type": "integer", "example": 23 },
          "daily_breakdown": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "date": { "type": "string", "format": "date" },
                "defect_count": { "type": "integer" },
                "soiling_count": { "type": "integer" },
                "efficiency_loss": { "type": "number" }
              }
            }
          },
          "trend": {
            "type": "object",
            "properties": {
              "soiling_trend": { "type": "string", "enum": ["increasing", "stable", "decreasing"], "example": "stable" },
              "recommendation": { "type": "string", "example": "현재 오염 수준은 정상 범위입니다." }
            }
          },
          "total_estimated_loss": { "type": "number", "example": 10500 }
        }
      },
      "EconomicReport": {
        "type": "object",
        "properties": {
          "panel_id": { "type": "integer", "example": 1 },
          "period": { "type": "string", "example": "month" },
          "analysis_date": { "type": "string", "format": "date", "example": "2026-01-16" },
          "soiling_impact": {
            "type": "object",
            "properties": {
              "average_soiling_coverage": { "type": "number", "example": 3.2, "description": "평균 오염 면적 (%)" },
              "estimated_efficiency_loss": { "type": "number", "example": 4.5, "description": "추정 효율 손실률 (%)" },
              "estimated_monthly_loss": { "type": "number", "example": 50000, "description": "추정 월 손실액 (원)" }
            }
          },
          "cleaning_recommendation": {
            "type": "object",
            "properties": {
              "is_cleaning_recommended": { "type": "boolean", "example": true },
              "estimated_cleaning_cost": { "type": "number", "example": 30000, "description": "청소 예상 비용 (원)" },
              "roi_if_cleaned": { "type": "number", "example": 20000, "description": "청소 시 예상 이익 (원)" },
              "recommendation_message": { 
                "type": "string", 
                "example": "현재 오염으로 인해 월 5만 원 손실 중. 청소비(3만 원)보다 손실이 크므로 청소를 권장합니다." 
              }
            }
          },
          "cost_benefit_analysis": {
            "type": "object",
            "properties": {
              "loss_without_cleaning": { "type": "number", "example": 150000, "description": "청소 미실시 시 3개월 예상 손실" },
              "cost_with_cleaning": { "type": "number", "example": 90000, "description": "청소 비용 (3회)" },
              "net_benefit": { "type": "number", "example": 60000, "description": "청소 실시 시 순이익" }
            }
          }
        }
      },
...
    }
  }
}